backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
cloudinary
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
